+++ common/types.go
@@ -49,7 +49,7 @@ import (
 const (
 	DefaultMaxLogFileSize = 512
 	DefaultLogFileCount   = 10
-	Blobfuse2Version      = "2.0.0-preview.3"
+	Blobfuse2Version      = "2.0.0-preview.4"
 	FileSystemName        = "blobfuse2"
 
 	DefaultConfigFilePath = "config.yaml"
@@ -61,12 +61,19 @@ const (
 	DefaultFilePermissionBits       os.FileMode = 0755
 	DefaultDirectoryPermissionBits  os.FileMode = 0775
 	DefaultAllowOtherPermissionBits os.FileMode = 0777
+
+	MbToBytes  = 1024 * 1024
+	BfuseStats = "blobfuse_stats"
 )
 
 var DefaultWorkDir = "$HOME/.blobfuse2"
 var DefaultLogFilePath = filepath.Join(DefaultWorkDir, "blobfuse2.log")
-var DefaultPipeline = []string{"libfuse", "file_cache", "attr_cache", "azstorage"}
-var DefaultStreamPipeline = []string{"libfuse", "stream", "attr_cache", "azstorage"}
+var StatsConfigFilePath = filepath.Join(DefaultWorkDir, "stats_monitor.cfg")
+
+var EnableMonitoring = false
+var BfsDisabled = false
+var TransferPipe = "/tmp/transferPipe"
+var PollingPipe = "/tmp/pollPipe"
 
 //LogLevel enum
 type LogLevel int
@@ -123,6 +130,7 @@ type LogConfig struct {
 	FileCount   uint64
 	FilePath    string
 	TimeTracker bool
+	Tag         string // logging tag which can be either blobfuse2 or bfusemon
 }
 
 // Flags for blocks