+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 **Features**
 - Added support for directory level SAS while mounting a subdirectory
 - Added support for displaying mount space utilization based on file cache consumption (for example when doing `df`)
+- Added support for updating MD5 sum on file upload
+- Added support for validating MD5 sum on download
+- Added backwards compatibility support for all blobfuse v1 CLI options
+- Added support to allow disabling writeback cache if a customer is opening a file with O_APPEND
+- Added support to ignore append flag on open when writeback cache is on
 
 **Bug Fixes**
 - Fixed a bug in parsing output of disk utilization summary